moving logout
